pyproject.toml
[tool.poetry]
name = "mystatus"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
jira = "^2.0.0"
PyYAML = "^5.4.1"
PyGithub = "^1.54.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
